+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - [Improvement] Allow for easier state usage by introducing explicit `#to_s` for reporting.
 - [Improvement] Change auto-generated id from `SecureRandom#uuid` to `SecureRandom#hex(6)`
 - [Improvement] Emit statistic every 5 seconds by default.
+- [Improvement] Introduce general messages parser that can be swapped when needed.
 - [Fix] Do not trigger code reloading when `consumer_persistence` is enabled.
 - [Fix] Shutdown producer after all the consumer components are down and the status is stopped. This will ensure, that any instrumentation related Kafka messaging can still operate.